requirements.txt
Pillow # import PIL

